requirements.txt
# Python Dependencies - Pandora Pipeline

# Core Scientific Computing
numpy>=1.26.4
pandas>=2.2.2
scipy>=1.11.0

# Image Processing
opencv-python>=4.10.0
pillow>=11.0.0
scikit-image>=0.22.0
pylsd-nova>=1.2.0  # Line Segment Detector (required for vanishing points)

# Geospatial
geopandas>=1.0.0
shapely>=2.0.0
pyproj>=3.5.0

# Cloud Storage
boto3>=1.34.0
botocore>=1.34.0
s3transfer>=0.10.0
laspy>=2.6.1

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.67.0

# Logging & Monitoring
coloredlogs>=15.0
psutil>=5.9.0

# Optional (per future ottimizzazioni GPU)
# torch>=2.0.0
# torchvision>=0.15.0
